+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Issue in `useFormStoreTable` where outdated local storage values caused columns to be hidden when they shouldn't be
+- attachment not uploaded if file is not stored locally successfully
 - Animations when scrolling from the validation errors toast
 
 ### Added